test: back part of ellipsis
