+++ components/typography/__tests__/index.test.js
@@ -3,7 +3,7 @@ import { mount } from 'enzyme';
 import { SmileOutlined, LikeOutlined, HighlightOutlined, CheckOutlined } from '@ant-design/icons';
 import KeyCode from 'rc-util/lib/KeyCode';
 import { resetWarned } from 'rc-util/lib/warning';
-import { spyElementPrototype, spyElementPrototypes } from 'rc-util/lib/test/domHook';
+import { spyElementPrototype } from 'rc-util/lib/test/domHook';
 import copy from 'copy-to-clipboard';
 import Title from '../Title';
 import Link from '../Link';
@@ -88,219 +88,6 @@ describe('Typography', () => {
   });
 
   describe('Base', () => {
-    describe('trigger ellipsis update', () => {
-      const fullStr =
-        'Bamboo is Little Light Bamboo is Little Light Bamboo is Little Light Bamboo is Little Light Bamboo is Little Light';
-
-      it('should trigger update', async () => {
-        const onEllipsis = jest.fn();
-        const wrapper = mount(
-          <Base ellipsis={{ onEllipsis }} component="p" editable>
-            {fullStr}
-          </Base>,
-        );
-
-        await sleep(20);
-        wrapper.update();
-        expect(wrapper.text()).toEqual('Bamboo is Little ...');
-        expect(onEllipsis).toHaveBeenCalledWith(true);
-        onEllipsis.mockReset();
-
-        wrapper.setProps({ ellipsis: { rows: 2, onEllipsis } });
-        await sleep(20);
-        wrapper.update();
-        expect(wrapper.text()).toEqual('Bamboo is Little Light Bamboo is Litt...');
-        expect(onEllipsis).not.toHaveBeenCalled();
-
-        wrapper.setProps({ ellipsis: { rows: 99, onEllipsis } });
-        await sleep(20);
-        wrapper.update();
-        expect(wrapper.find('p').text()).toEqual(fullStr);
-        expect(onEllipsis).toHaveBeenCalledWith(false);
-
-        wrapper.unmount();
-      });
-
-      it('string with parentheses', async () => {
-        const parenthesesStr = `Ant Design, a design language (for background applications, is refined by
-          Ant UED Team. Ant Design, a design language for background applications,
-          is refined by Ant UED Team. Ant Design, a design language for background
-          applications, is refined by Ant UED Team. Ant Design, a design language
-          for background applications, is refined by Ant UED Team. Ant Design, a
-          design language for background applications, is refined by Ant UED Team.
-          Ant Design, a design language for background applications, is refined by
-          Ant UED Team.`;
-        const onEllipsis = jest.fn();
-        const wrapper = mount(
-          <Base ellipsis={{ onEllipsis }} component="p" editable>
-            {parenthesesStr}
-          </Base>,
-        );
-
-        await sleep(20);
-        wrapper.update();
-        expect(wrapper.text()).toEqual('Ant Design, a des...');
-        const ellipsisSpan = wrapper.find('span[title]');
-        expect(ellipsisSpan.text()).toEqual('...');
-        expect(ellipsisSpan.props().title)
-          .toEqual(`ign language (for background applications, is refined by
-          Ant UED Team. Ant Design, a design language for background applications,
-          is refined by Ant UED Team. Ant Design, a design language for background
-          applications, is refined by Ant UED Team. Ant Design, a design language
-          for background applications, is refined by Ant UED Team. Ant Design, a
-          design language for background applications, is refined by Ant UED Team.
-          Ant Design, a design language for background applications, is refined by
-          Ant UED Team.`);
-        onEllipsis.mockReset();
-
-        wrapper.unmount();
-      });
-
-      it('should middle ellipsis', async () => {
-        const suffix = '--suffix';
-        const wrapper = mount(
-          <Base ellipsis={{ rows: 1, suffix }} component="p">
-            {fullStr}
-          </Base>,
-        );
-
-        await sleep(20);
-        wrapper.update();
-        expect(wrapper.find('p').text()).toEqual('Bamboo is...--suffix');
-        wrapper.unmount();
-      });
-
-      it('should front or middle ellipsis', async () => {
-        const suffix = '--The information is very important';
-        const wrapper = mount(
-          <Base ellipsis={{ rows: 1, suffix }} component="p">
-            {fullStr}
-          </Base>,
-        );
-
-        await sleep(20);
-        wrapper.update();
-        expect(wrapper.find('p').text()).toEqual('...--The information is very important');
-
-        wrapper.setProps({ ellipsis: { rows: 2, suffix } });
-        await sleep(20);
-        wrapper.update();
-        expect(wrapper.find('p').text()).toEqual('Ba...--The information is very important');
-
-        wrapper.setProps({ ellipsis: { rows: 99, suffix } });
-        await sleep(20);
-        wrapper.update();
-        expect(wrapper.find('p').text()).toEqual(fullStr + suffix);
-
-        wrapper.unmount();
-      });
-
-      it('connect children', async () => {
-        const bamboo = 'Bamboo';
-        const is = ' is ';
-
-        const wrapper = mount(
-          <Base ellipsis component="p" editable>
-            {bamboo}
-            {is}
-            <code>Little</code>
-            <code>Light</code>
-          </Base>,
-        );
-
-        await sleep(20);
-        wrapper.update();
-
-        expect(wrapper.text()).toEqual('Bamboo is Little...');
-      });
-
-      it('should expandable work', async () => {
-        const onExpand = jest.fn();
-        const wrapper = mount(
-          <Base ellipsis={{ expandable: true, onExpand }} component="p" copyable editable>
-            {fullStr}
-          </Base>,
-        );
-
-        await sleep(20);
-        wrapper.update();
-
-        wrapper.find('.ant-typography-expand').simulate('click');
-        expect(onExpand).toHaveBeenCalled();
-        await sleep(20);
-        wrapper.update();
-
-        expect(wrapper.find('p').text()).toEqual(fullStr);
-      });
-
-      it('should have custom expand style', async () => {
-        const symbol = 'more';
-        const wrapper = mount(
-          <Base ellipsis={{ expandable: true, symbol }} component="p">
-            {fullStr}
-          </Base>,
-        );
-        await sleep(20);
-        wrapper.update();
-        expect(wrapper.find('.ant-typography-expand').text()).toEqual('more');
-      });
-
-      it('can use css ellipsis', () => {
-        const wrapper = mount(<Base ellipsis component="p" />);
-        expect(wrapper.find('.ant-typography-ellipsis-single-line').length).toBeTruthy();
-      });
-
-      it('should calculate padding', () => {
-        const wrapper = mount(
-          <Base ellipsis component="p" style={{ paddingTop: '12px', paddingBottom: '12px' }} />,
-        );
-        expect(wrapper.find('.ant-typography-ellipsis-single-line').length).toBeTruthy();
-      });
-
-      describe('should tooltip support', () => {
-        let domSpy;
-
-        beforeAll(() => {
-          domSpy = spyElementPrototypes(HTMLElement, {
-            offsetWidth: {
-              get: () => 100,
-            },
-            scrollWidth: {
-              get: () => 200,
-            },
-          });
-        });
-
-        afterAll(() => {
-          domSpy.mockRestore();
-        });
-
-        function getWrapper(tooltip) {
-          return mount(
-            <Base ellipsis={{ tooltip }} component="p">
-              {fullStr}
-            </Base>,
-          );
-        }
-
-        it('boolean', async () => {
-          const wrapper = getWrapper(true);
-          await sleep(20);
-          wrapper.update();
-
-          expect(wrapper.find('Tooltip').prop('title')).toEqual(fullStr);
-        });
-
-        it('customize', async () => {
-          const wrapper = getWrapper('Bamboo is Light');
-          await sleep(20);
-          wrapper.update();
-
-          expect(wrapper.find('Tooltip').prop('title')).toEqual('Bamboo is Light');
-        });
-      });
-    });
-
     describe('copyable', () => {
       function copyTest(name, text, target, icon, tooltips) {
         it(name, async () => {
